requirements.txt
fastapi[all]~=0.112.0
psycopg2~=2.9.9
#passlib[bcrypt]~=1.7.4
sqlalchemy~=2.0.32
pydantic~=2.8.2
#pyjwt~=2.9.0
uvicorn~=0.30.6
requests~=2.32.3
pillow~=10.4.0
pydantic-settings~=2.4.0